Use try pattern for OS parsing
