+++ src/UADetector/Regexes/Models/IRegexDefinition.cs
@@ -2,7 +2,7 @@ using System.Text.RegularExpressions;
 
 namespace UADetector.Regexes.Models;
 
-public interface IRegexDefinition
+internal interface IRegexDefinition
 {
     public Regex Regex { get; init; }
     public string Name { get; init; }
