+++ src/UADetector/Regexes/Models/Os.cs
@@ -2,7 +2,7 @@ using System.Text.RegularExpressions;
 
 namespace UADetector.Regexes.Models;
 
-public sealed class Os : IRegexDefinition
+internal sealed class Os : IRegexDefinition
 {
     public required Regex Regex { get; init; }
     public required string Name { get; init; }
