+++ src/UADetector/Results/UserAgentInfo.cs
@@ -2,5 +2,5 @@ namespace UADetector.Results;
 
 public class UserAgentInfo
 {
-
+    public OsInfo? Os { get; internal set; }
 }
